+++ packages/dev-server/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @cosmicdrift/kumiko-dev-server
 
+## 0.245.0
+
+### Patch Changes
+
+- Updated dependencies [3359dae]
+- Updated dependencies [d669ad6]
+  - @cosmicdrift/kumiko-framework@0.245.0
+  - @cosmicdrift/kumiko-bundled-features@0.245.0
+  - @cosmicdrift/kumiko-server-runtime@0.245.0
+
 ## 0.244.0
 
 ### Patch Changes